Move from @/inject/apis to @/inject/models
